Add exercises up to incl 2.18. Need refactor.

+++ part2/phonebook/src/App.js
@@ -19,21 +19,41 @@ const App = () => {
             .getAll()
             .then(currentPersons => setPersons(currentPersons));
     };
-
     useEffect(getPersons, []);
 
     const addPerson = event => {
         event.preventDefault();
 
-        if (persons.filter(e => e.name === newName).length > 0) {
-            setNewName('');
-            return alert(`${newName} is already in the phonebook!`);
-        }
-
         const personObj = {
             name: newName,
             number: newNumber,
         };
+
+        if (persons.filter(person => person.name === newName).length > 0) {
+            setNewName('');
+            setNewNumber('');
+            const result = window.confirm(
+                `${newName} is already in the phonebook! Update phone number?`
+            );
+            if (result === true) {
+                const person = persons.find(person => person.name === newName);
+                const changedPerson = { ...person, number: newNumber };
+                personService
+                    .update(changedPerson.id, changedPerson)
+                    .then(returnedPerson => {
+                        console.log(returnedPerson);
+                        console.log(persons);
+                        setPersons(
+                            persons.map(person =>
+                                person.id !== returnedPerson.id
+                                    ? person
+                                    : returnedPerson
+                            )
+                        );
+                    });
+            }
+            return null;
+        }
 
         personService.post(personObj).then(returnedPerson => {
             setPersons(persons.concat(returnedPerson));
@@ -43,9 +63,12 @@ const App = () => {
     };
 
     const deletePerson = id => {
-        personService.destroy(id).then(() => {
-            setPersons(persons.filter(person => person.id !== id));
-        });
+        const name = persons.find(person => person.id === id).name;
+        if (window.confirm(`Delete ${name}?`)) {
+            personService.destroy(id).then(() => {
+                setPersons(persons.filter(person => person.id !== id));
+            });
+        }
     };
 
     const handleNameChange = event => {
